Support cert-manager, fix apiserver build dependencies (#329)
* Add missing dependencies, pin etcd version

* Add version.go to .gitignore (build set this value anyways)

* Add support for certmgr config

* Synchronize envvar name

* Added some sanity checks

* Fix bad varname

* One more sanity check

* Remove underscore, use camel case in apiserver.json

* Add default values to apiserver.json

* Fixed TLS host lists for started applications

* Update to Golang 1.15, slimmed down and sped up Docker build

* Add rootHost alongside wildcard in Spec.Ingress.TLS

* Don't add secretName (reuse secret in default namespace)

+++ build.sh
@@ -20,7 +20,7 @@ if [ "$1" == "local" ] || [ "$1" == "docker" ]; then
     # Check for --cache flag
     args="$@"
     replaced="${@/--cache/}"
-    if [ "$args" == "$replaced" ]; then
+    if [ "$1" == "local" ] && [ "$args" == "$replaced" ]; then
         echo "Fetching dependencies..."
         glide install --strip-vendor
     fi
